Refactor out type_check
type_check and related functions don't need to be blanket implementations of
the Property trait, they can be simple impls. This simplify things and
allows to remove 2 unreachable!()
